requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
qdrant-client==1.17.0
# sentence-transformers  # optional: only needed for local BGE embeddings
httpx==0.28.1
pypdf==6.7.5
python-multipart==0.0.22
langgraph==1.0.10
langchain-core==1.2.17
sentry-sdk==2.54.0
prometheus-client==0.24.1
python-dateutil==2.9.0.post0
sqlalchemy==2.0.48
alembic==1.18.4
psycopg[binary]==3.3.3
redis==7.3.0
rq==2.7.0
openai==2.26.0
langsmith==0.7.14
